Begin writing tests
tests for frog and lilypad mostly finished
