+++ index.html
@@ -9,7 +9,6 @@
   <div class="game-container" id="game-container">
     <div class="btn-container">
       <button id="start-button">Start</button>
-      <button id="reset-button">Reset</button>
       <button id="pause-button">Pause</button>
       <button id="new-game-button">New Game</button>
       <h1 id="lives">Lives: <span class="lives-left">3</span></h1>
