+++ lib/Frog.js
@@ -12,8 +12,6 @@ class Frog {
     this.isAlive = true;
     this.lives = 3;
     this.deathClock = 0;
-    this.paused = false;
-    this.gameOver = false;
   }
 
   draw(c) {
@@ -22,27 +20,23 @@ class Frog {
     return this;
   }
 
-  move(canvas) {
-    window.addEventListener( 'keydown', (event) => {
-      event.preventDefault();
-      if (this.paused === false && this.isAlive === true) {
+  move(canvas, pressedButton) {
+    if (this.isAlive === true) {
       this.deathClock = 0;
-        if (event.key === 'ArrowLeft' &&
-          this.x - this.width * 1.5 > 0) {
-          this.moveLeft();
-        } else if (event.key === 'ArrowRight' &&
-          this.x + this.width * 2.5 < canvas.width) {
-          this.moveRight();
-        } else if (event.key === 'ArrowUp' &&
-          this.y >= 126) {
-          this.moveUp();
-        } else if (event.key === 'ArrowDown' &&
-          this.y < 500) {
-          this.moveDown();
-        }
+      if (pressedButton === 'ArrowLeft' &&
+        this.x - this.width * 1.5 > 0) {
+        this.moveLeft();
+      } else if (pressedButton === 'ArrowRight' &&
+        this.x + this.width * 2.5 < canvas) {
+        this.moveRight();
+      } else if (pressedButton === 'ArrowUp' &&
+        this.y >= 126) {
+        this.moveUp();
+      } else if (pressedButton === 'ArrowDown' &&
+        this.y < 500) {
+        this.moveDown();
       }
-      return this;
-    })
+    }
   }
 
   moveLeft() {
@@ -69,12 +63,12 @@ class Frog {
           this.x + this.width > car.x &&
           this.y < car.y + car.height &&
           this.height + this.y > car.y) === true) {
-        this.isAlive = false;
+        return this.isAlive = false;
       }
     })
   }
 
-  onALog (logsLeft, turtlesRight, water) {
+  floatOrSink (logsLeft, turtlesRight, water) {
     var allLogs = [...logsLeft, ...turtlesRight];
     if (!this.isOnLog(allLogs) && this.isInWater(water)) {
       this.isAlive = false;
@@ -82,40 +76,35 @@ class Frog {
   }
 
   isOnLog(allLogs) {
-    let outPutVal = false;
+    let onALog = false;
     
     allLogs.forEach(log => {
       if (this.x < log.x + log.width - this.width / 2 &&
              this.x + this.width / 2 > log.x &&
              this.y < log.y + log.height &&
              this.height + this.y > log.y) {
-        outPutVal = true;
+        onALog = true;
         this.x += log.dx;
       }
     })
-    return outPutVal;
+    return onALog;
   }
 
   isInWater(water) {
-    let outPutVal = false;
-    
-    if (this.x < water.x + water.width - this.width / 2 &&
-        this.x + this.width / 2 > water.x &&
-        this.y < water.y + water.height &&
-        this.height + this.y > water.y) {
-      outPutVal = true;
-    }
-    return outPutVal;
+    return this.x < water.x + water.width - this.width / 2 &&
+           this.x + this.width / 2 > water.x &&
+           this.y < water.y + water.height &&
+           this.height + this.y > water.y;
   }
 
-  onLilyPad (lilyArray, canvas) {
+  onLilyPad (lilyArray) {
     if (this.landedOnPad(lilyArray)) {
-      this.respawn(canvas);
+      this.respawn();
     }
   }
 
   landedOnPad (lilyArray) {
-    let outPutVal = false;
+    let onLilyPad = false;
     
     lilyArray.forEach(lilyPad => {
       if ((this.x < lilyPad.x + lilyPad.width &&
@@ -123,11 +112,12 @@ class Frog {
            this.y < lilyPad.y + lilyPad.height &&
            this.height + this.y > lilyPad.y) === true &&
            lilyPad.color !== 'lightgreen') {
-        outPutVal = true;
+        onLilyPad = true;
         lilyPad.landedOn = true;
+        lilyPad.color = 'lightgreen';
       } 
     })
-    return outPutVal;
+    return onLilyPad;
   }
 
   outOfBounds(canvas) {
@@ -137,7 +127,7 @@ class Frog {
     }
   }
 
-  die(canvas) {
+  die() {
     if (this.isAlive === false) {
       this.color = 'red';
       this.deathClock++;
@@ -152,21 +142,18 @@ class Frog {
       } else {
         this.lives--;
         this.deathClock = 0;
-        this.respawn(canvas);
+        this.respawn();
       }
     }
   }
 
-  respawn(canvas) {
+  respawn() {
+    var frogStartPoint = { x : 190, y : 500 };
     if (this.lives >= 0) {
-      this.x = canvas.width / 2 - 10;
-      this.y = 500;
+      this.x = frogStartPoint.x;
+      this.y = frogStartPoint.y;
       this.isAlive = true;
       this.color = 'lightgreen';
-    } else {
-      this.lives = 0;
-      console.log('game over')
-      this.gameOver = true;
     }
   }
 }
