+++ lib/LilyPad.js
@@ -14,7 +14,6 @@ class LilyPad {
     }
     c.fillStyle = this.color;
     c.fillRect(this.x, this.y, this.width, this.height);
-
   }
 }
