+++ test/Frog-test.js
@@ -1,13 +1,23 @@
 const {assert} = require('chai');
 const Frog = require('../lib/Frog.js');
+const Vehicle = require('../lib/Vehicle.js');
+const Water = require('../lib/Water.js')
+const Log = require('../lib/Log.js')
+const LilyPad = require('../lib/LilyPad.js')
+
 
 describe('Frog', function() {
+  let frogger;
+
+  beforeEach( function() {
+    frogger = new Frog(10,10,10,10,10,10,'green');
+  });
+
   it('Should be a function', function() {
     assert.isFunction(Frog);
   })
  
   it('Should be able to make a new frogger', function() {
-    var frogger = new Frog();
     assert.isObject(frogger);
   })
   
@@ -24,8 +34,6 @@ describe('Frog', function() {
   })
   
   it('Should be able to move up', function() {
-    var frogger = new Frog(10,10,10,10,10,10,'green');
-
     assert.equal(frogger.y, 10);
 
     frogger.moveUp();
@@ -34,8 +42,6 @@ describe('Frog', function() {
   })
   
   it('Should be able to move down', function() {
-    var frogger = new Frog(10,10,10,10,10,10,'green');
-
     assert.equal(frogger.y, 10);
 
     frogger.moveDown();
@@ -44,8 +50,6 @@ describe('Frog', function() {
   })
   
   it('Should be able to move left', function() {
-    var frogger = new Frog(10,10,10,10,10,10,'green');
-
     assert.equal(frogger.x, 10);
 
     frogger.moveLeft();
@@ -54,8 +58,6 @@ describe('Frog', function() {
   }) 
 
   it('Should be able to move right', function() {
-    var frogger = new Frog(10,10,10,10,10,10,'green');
-
     assert.equal(frogger.x, 10);
 
     frogger.moveRight();
@@ -63,13 +65,93 @@ describe('Frog', function() {
     assert.equal(frogger.x, 20);
   })
 
-  // Should be able to stand on turtle/log
-  // Should move the same speed as turtle/log while on it.
-  // Should respawn if killed
-  // Should respawn if landed on lilypad
-  // Should be killed by car
-  // Should be killed by water
-  // Should lose life if killed
-  // Should report 'game over' if all lives are gone
+  it('Should die when colliding with a car', function() {
+    var leftCar = new Vehicle(10,10,10,10,1,'red');
+    var rightCar = new Vehicle(20,20,20,20,1,'red');
 
+    assert.equal(frogger.isAlive, true)
+
+    frogger.hitByCar([leftCar], [rightCar])
+
+    assert.equal(frogger.isAlive, false);
+  })
+
+  it('Should die when on water and not on a log', function () {
+    let water = new Water(10,10,10,10)
+    let leftLog = new Log(20,20,20,20)
+    let rightLog = new Log(30,30,30,30)
+
+    assert.equal(frogger.isAlive, true)
+
+    frogger.floatOrSink([leftLog], [rightLog],water)
+
+    assert.equal(frogger.isAlive, false)
+  })
+
+  it('Should move with a log', function() {
+    let log = new Log(10,10,10,10,10,'brown')
+
+    assert.equal(frogger.x, 10)
+    frogger.isOnLog([log])
+    assert.equal(frogger.x, 20)
+  })
+
+  it('Should not die when on a log or turtle', function() {
+    let water = new Water(10,10,10,10)
+    let leftLog = new Log(10,10,10,10)
+    let rightTurtle = new Log(30,30,30,30)
+
+    assert.equal(frogger.isAlive, true)
+
+    frogger.floatOrSink([leftLog], [rightTurtle],water)
+
+    assert.equal(frogger.isAlive, true)
+  })
+
+  it('Should lose a life if it is killed', function() {
+    assert.equal(frogger.lives, 3)
+    
+    assert.equal(frogger.isAlive, true)
+    
+    frogger.isAlive = false;
+    frogger.deathClock = 80;
+    
+    assert.equal(frogger.isAlive, false)
+    
+    frogger.die()
+
+    assert.equal(frogger.lives, 2)
+  })
+
+  it('Should go back to start when it respawns', function() {
+    assert.equal(frogger.x, 10)
+
+    frogger.moveLeft()
+
+    assert.equal(frogger.x, 0)
+
+    frogger.respawn()
+
+    assert.equal(frogger.x, 190);
+  })
+
+  it('Should respawn when it dies', function() {
+    assert.equal(frogger.x, 10)
+
+    frogger.isAlive = false;
+    frogger.deathClock = 80;
+    frogger.die();
+
+    assert.equal(frogger.x, 190)
+  })
+
+  it('Should respawn after landing on a lilyPad', function() {
+    let lilyPad = new LilyPad(10,10,10,10)
+
+    assert.equal(frogger.x, 10)
+
+    frogger.onLilyPad([lilyPad])
+
+    assert.equal(frogger.x, 190)
+  })
 })
