+++ test/Turtle-test.js
@@ -1,5 +1,5 @@
 // Should be function
 // Should be able to make an instance of itself
 // Should be able to take paramaters
-// should be able to carry frog
 // should be able to move left/right
+// Should increase speed on the next level
