+++ test/Vehicle-test.js
@@ -2,4 +2,4 @@
 // Should be able to make an instance of itself
 // Should be able to take paramaters
 // Should be able to move left/right
-// Should kill frog if they collide
+// Should increase speeed on next level
