+++ test/Water-test.js
@@ -1,4 +1,3 @@
 // Should be function
 // Should be able to make an instance of itself
 // Should be able to take paramaters
-// Should kill frog if he steps on it
